First stage in dndMovement, I've allowed keyboard control of a single token.

+++ Open_Games_System/src/SnakeLadderGrid.java
@@ -8,17 +8,17 @@ public class SnakeLadderGrid extends Grid{
     }
 
     public SnakeLadderGrid(){}
-    public SnakeLadderGrid(int x, int y, int height, int width){
+    public SnakeLadderGrid(int x, int y, double height, double width){
         this.x = x;
         this.y = y;
         tiles = new Tile[x*y];
         populateTiles(height,width);
     }
     @Override
-    public void populateTiles(int height, int width) {
-        int inc = width/x;
-        int n = - inc; //starting point of x at the left
-        int m = height - height/y; //last point at y - 1 tile
+    public void populateTiles(double height, double width) {
+        double inc = width/x;
+        double n = - inc; //starting point of x at the left
+        double m = height - height/y; //last point at y - 1 tile
         int count = 0; // will always be x * y
         /*for (int i = 0; i<x ; i++,n+= width/x, m=0)
             for (int j = 0; j<y ; j++,m += height/y )
@@ -29,7 +29,7 @@ public class SnakeLadderGrid extends Grid{
                     n += inc;
                 }
                 else toggle = !toggle;
-                tiles[count] = new Tile(n, m, count++);
+                tiles[count] = new Tile((int)n, (int)m, count++);
 
             }
             toggle = !toggle;
